Added telegram bot
Now through a chatbot in telegram, we can send post request to the API and receive immediate response on the telegram itself.

One thing could be done more, that we send text from telegram, received by our application frontend which then makes API calls but it requires webhooks setup. We are not going into that.